NON-145: Ability to load existing data into a form
Without validating it. I've moved this operation into its own Form method
for clarity.

This is useful when for example a form is for editing existing data which
may or may not be valid. Validate this data would mean the user would
see an error even before interacting with the form (load and Validate
is still possibly by using `validate()`).