NON-169: Creating cypress tests - WIP (#86)
* NON-169: Set up integration test file

* Adding NonAssociation test pages. WIP

* Nearly there?

* Fix non-associations api stubs for integrations tests: now the endpoints have the correct prefix BUT ignore query parameters

* minor updates to cypress tests

* Successful add prisonerNonAssociation test.

* Successful close prisonerNonAssociation test.

* cleaning up add/close integration tests

* Update all non major NPM dependencies (#73)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Expose product ID (#69)

* Update all non major NPM dependencies to ^6.3.0 (#74)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Fix typo (#75)

* Update dependency lint-staged to v14 (#78)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update all non major NPM dependencies (#77)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update dependency prettier to v3 (#23)

* Update dependency eslint-plugin-prettier to v5
* Update dependency prettier to v3
* Update dependency prettier-plugin-jinja-template to v1
* Appease eslint

* Update dependency eslint to ^8.47.0 (#79)

* Update dependency @types/node to ^18.17.5 (#80)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update all non major NPM dependencies (#82)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update all non major NPM dependencies (#83)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update all non major NPM dependencies (#84)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* NON-152/NON-145/NON-145: Ability to edit a non-association (#85)

* NON-152: Link to update a non-association

* NON-145: Improved tests for update route

* NON-145: Added UpdateForm

* NON-145: Moved non-association form/page into a partial and reuse it

Pages to add and edit a non-association are identical.

* NON-145: Updated update route to update non-association

* NON-145: Moved add/update form validation in common parent form

The validations for creating a new non-association are still relevant
when updating a non-association so it makes sense to have these in a
common parent class.

Updating a non-association has an additional validation (comment should
change) that can go in the specific update form.

* NON-145: Allow population of form without validation

This is useful in the case of the 'update' form because one of the
validation rules is to force user to change the non-association comment.

However in order to load the existing non-association data into the form
and show it to the user the route has to "submit" the form with the
existing data. This triggered the validation which in turn caused meant
user would see an error on the `comment` field even before they actually
interacted with the form.

Now it's possible to submit/load data into the form but without triggering
the validation.

* NON-145: Added test for loading non-association data on update

* NON-145: More tests for the update route

* Use `extend` in non-association add/update templates

* NON-145: Ability to load existing data into a form

Without validating it. I've moved this operation into its own Form method
for clarity.

This is useful when for example a form is for editing existing data which
may or may not be valid. Validate this data would mean the user would
see an error even before interacting with the form (load and Validate
is still possibly by using `validate()`).

* NON-152: Removed debug statements

* lint fix

* NON-152: Ensures that constants are preserved when mocking

As per PR comment

* Update all non major NPM dependencies (#87)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* NON-145: Prevent editing of closed non-associations (#88)

Once non-associations are closed can't be edited directly by the user.

Route returns 404 to indicate the page to edit a closed non-association
doesn't "exist".

* Removed TODO

* Refactoring names of classes

---------

Co-authored-by: Igor <igor.ushkarev@digital.justice.gov.uk>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: ushkarev <ushkarev@users.noreply.github.com>
Co-authored-by: Aldo Giambelluca <xoen@users.noreply.github.com>