NON-209: Track clicks on confirmation pages
The Add/Update/Close confirmation pages have a link that takes the user
back to the key prisoner's non-associations page.

Clicking on these now publishes a custom GA event (name is `non_associations_event`).

The `category` field in the event can be used to filter/discriminate between
pages.

In this specific case `action`/`label` fields are blank but the mechanism
that publish these events will potentially include these if the
`data-ga-action`/`data-ga-label` attributes are set (but in the simple
case of clicks these are not necessary).

NOTE: The `active_case_load` property is automatically set as for other
events published, even if this is not passed explicitly on send.