NON-394: Don't look up `authorisedBy` "users"
This field contains valid usernames now but in the
past was free text meaning that we can't guarantee
it contains valid usernames.

For invalid usernames the Prison API requests will
return a `404 Not Found`, this is handled fine but
it doesn't change the fact that this is just wasteful.

It also seems like the UI doesn't use/display these
"authorised by" fields at all, making this lookup
even more wasteful.